% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gwa.internal}
\alias{gwa.internal}
\title{Simple GWA}
\usage{
gwa.internal(yVec, kVec, snps, d, Ngen, nThr)
}
\arguments{
\item{yVec}{vectorized matrix of phenotypes}

\item{kVec}{vectorized relationship matrix}

\item{snps}{SNP matrix, SNPs as columns}

\item{d}{number of traits}

\item{Ngen}{number of genotypes}

\item{nThr}{number of threads}
}
\description{
Fits a random-effects model (with no fixed effect covariates other than the intercept and no replicated measurement of genotypes) and does GWA on the provided SNPs. Operates on any number of traits at once, but treats them as independent. If the number of threads is set to zero, all available cores are used.
}
