% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaceFunctions.R
\name{gwa}
\alias{gwa}
\title{Single-marker genome-wide association}
\usage{
gwa(Y, K, X = NULL, repFactor = NULL, snps, nPerm = NULL,
  nThreads = NULL)
}
\arguments{
\item{Y}{matrix of phenotypes; phenotypes should be in columns.}

\item{K}{relationship matrix}

\item{X}{fixed effect matrix (optional, do not include the intercept)}

\item{repFactor}{factor relating genotypes to replicates (optional)}

\item{snps}{matrix of genotypes (\code{NA} for missing data, lines as rows, SNPs as columns)}

\item{nPerm}{number of permutations to calculate \eqn{q}-values}

\item{nThreads}{suggested number of threads; if \code{NULL} or 0, maximum number will be used}
}
\value{
a list containing a matrix of random effects (named \code{ranef}), a matrix of fixed effects (including the intercept, named \code{fixef}), a vector of heritability estimates (named \code{hSq}), matrix of \eqn{-\log_{10}p} values (named \code{lPval}), and an optinal matrix of \eqn{q}-values (named \code{qVal}).
}
\description{
Performs single-marker GWA with a mixed-model population structure correction. The algorithm is EMMAX. Optionally supports replicated genotypes and fixed-effect covariates. Empirical \eqn{q}-values can be calculated using permutations is desired. Since all loci are used in the permutation test, the number of samples need not be very large. I get good results with 15 to 25 permutations. Phenotype, relationship matrix, covariate, and replication specification is the same as in \code{\link{mmFit}}. The genotype matrix may have missing data (specified as \code{NA}). Genotypes should be coded as {0, 1, 2}, with 1 as the heterozygote. The lines should be in rows, with the number of rows equal to the number of lines.
}
\details{
It is possible to specify the number of cores to be used. Multithreading is employed only for the SNP effect estimation. If you compiled R with multithreaded BLAS and/or LAPACK, it is safe to use it with this function. The linear algebra functions are run on single threads. The specified number of threads will be taken as a limit. If there are too few genotypes, not all of the suggested cores may be deployed. If the number of threads is \code{NULL}, 0, or not specified, the maximum number of cores will be used, provided there are enough SNPs to do so.
}
\references{
\insertRef{kang08}{GWAlikeMeth}
\insertRef{kang10}{GWAlikeMeth}
}
