% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gwaR.internal}
\alias{gwaR.internal}
\title{GWA with replication}
\usage{
gwaR.internal(yVec, kVec, repFac, snps, d, Ngen, nThr)
}
\arguments{
\item{yVec}{vectorized matrix of phenotypes}

\item{kVec}{vectorized relationship matrix}

\item{repFac}{factor relating genotypes to replicates}

\item{snps}{SNP matrix, SNPs as columns}

\item{d}{number of traits}

\item{Ngen}{number of genotypes}

\item{nThr}{number of threads}
}
\description{
Fits a random-effects model (with no fixed effect covariates other than the intercept but with genotypes measured in replicates) and does GWA on the provided SNPs. Operates on any number of traits at once, but treats them as independent. All available cores will be used if the number of threads is set to zero.
}
