% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gwaRF.internal}
\alias{gwaRF.internal}
\title{GWA with replication and fixed effects}
\usage{
gwaRF.internal(yVec, kVec, repFac, xVec, snps, d, Ngen, nThr)
}
\arguments{
\item{yVec}{vectorized matrix of phenotypes}

\item{kVec}{vectorized relationship matrix}

\item{repFac}{factor relating genotypes to replicates}

\item{xVec}{vectorized relationship matrix}

\item{snps}{SNP matrix, SNPs as columns}

\item{d}{number of traits}

\item{Ngen}{number of genotypes}

\item{nThr}{number of threads}
}
\description{
Fits a random-effects model (with fixed effect covariates and genotypes measured in replicates) and does GWA on the provided SNPs. Operates on any number of traits at once, but treats them as independent. All available cores will be used if the number of threads is set to zero.
}
