% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reFit}
\alias{reFit}
\title{Random effects fit}
\usage{
reFit(yVec, kVec, d, Ngen)
}
\arguments{
\item{yVec}{vectorized matrix of phenotypes}

\item{kVec}{vectorized relationship matrix}

\item{d}{number of traits}

\item{Ngen}{number of genotypes}
}
\description{
Fits a random-effects model (with no fixed effect covariates other than the intercept and no replication). Operates on any number of traits at once, but treats them as independent.
}
