% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reFitF}
\alias{reFitF}
\title{Random effects fit with fixed effects but no replication}
\usage{
reFitF(yVec, kVec, xVec, d, Ngen)
}
\arguments{
\item{yVec}{vectorized matrix of phenotypes}

\item{kVec}{vectorized relationship matrix}

\item{xVec}{vectorized matrix of fixed effects}

\item{d}{number of traits}

\item{Ngen}{number of genotypes}
}
\description{
Fits a random-effects model (with fixed effect covariates). Operates on any number of traits at once, but treats them as independent.
}
