% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reFitR}
\alias{reFitR}
\title{Random effects fit with replication}
\usage{
reFitR(yVec, kVec, repFac, d, Ngen)
}
\arguments{
\item{yVec}{vectorized matrix of phenotypes}

\item{kVec}{vectorized relationship matrix}

\item{repFac}{factor relating genotypes to replicates}

\item{d}{number of traits}

\item{Ngen}{number of genotypes}
}
\description{
Fits a random-effects model (with no fixed effect covariates other than the intercept). Operates on any number of traits at once, but treats them as independent.
}
