% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reFitRF}
\alias{reFitRF}
\title{Random effects fit with fixed effects and replication}
\usage{
reFitRF(yVec, kVec, repFac, xVec, d, Ngen)
}
\arguments{
\item{yVec}{vectorized matrix of phenotypes}

\item{kVec}{vectorized relationship matrix}

\item{repFac}{factor relating genotypes to replicates}

\item{xVec}{vectorized matrix of fixed effects}

\item{d}{number of traits}

\item{Ngen}{number of genotypes}
}
\description{
Fits a random-effects model (with fixed effect covariates and replicated genotype measurements). Operates on any number of traits at once, but treats them as independent.
}
